requirements.txt
pandas
numpy
matplotlib
scipy
yfinance
sqlalchemy
mysql-connector-python
streamlit
openpyxl